+++ src/main/java/org/mmarini/wheelly/engines/AvoidingState.java
@@ -31,8 +31,18 @@ package org.mmarini.wheelly.engines;
 import com.fasterxml.jackson.databind.JsonNode;
 import org.mmarini.wheelly.apis.RobotStatus;
 import org.mmarini.yaml.schema.Locator;
+import org.mmarini.yaml.schema.Validator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.awt.geom.Point2D;
+import java.util.List;
+import java.util.Map;
+import java.util.OptionalInt;
+
+import static java.lang.String.format;
+import static org.mmarini.wheelly.apis.Utils.normalizeDegAngle;
+import static org.mmarini.yaml.schema.Validator.positiveNumber;
 
 /**
  * Generates the behavior to avoid contact obstacle.
@@ -43,16 +53,38 @@ import org.slf4j.LoggerFactory;
  * </p>
  */
 public class AvoidingState extends AbstractStateNode {
+    public static final String FREE_POINT = "freePoint";
+    public static final String SAFE_DISTANCE = "safeDistance";
+    public static final String ESCAPE_DIRECTION = "escapeDirection";
+    public static final Validator STATE_SPEC = Validator.objectPropertiesRequired(Map.of(
+                    "safeDistance", positiveNumber()
+            ), List.of(
+                    "safeDistance")
+    );
+
     private static final Logger logger = LoggerFactory.getLogger(AvoidingState.class);
 
     public static AvoidingState create(JsonNode root, Locator locator, String id) {
         BASE_STATE_SPEC.apply(locator).accept(root);
+        STATE_SPEC.apply(locator).accept(root);
         ProcessorCommand onInit = ProcessorCommand.concat(
                 loadTimeout(root, locator, id),
+                ProcessorCommand.put(id + "." + SAFE_DISTANCE, locator.path(SAFE_DISTANCE).getNode(root).doubleValue()),
                 ProcessorCommand.create(root, locator.path("onInit")));
         ProcessorCommand onEntry = ProcessorCommand.create(root, locator.path("onEntry"));
         ProcessorCommand onExit = ProcessorCommand.create(root, locator.path("onExit"));
+
         return new AvoidingState(id, onInit, onEntry, onExit);
+    }
+
+    static OptionalInt escapeDir(ProcessorContext context) {
+        RobotStatus status = context.getRobotStatus();
+        int robotDir = status.getDirection();
+        return !status.canMoveForward() ?
+                status.canMoveBackward() ?
+                        OptionalInt.of(normalizeDegAngle(robotDir + 180))
+                        : OptionalInt.empty()
+                : OptionalInt.of(robotDir);
     }
 
     protected AvoidingState(String id, ProcessorCommand onInit, ProcessorCommand onEntry, ProcessorCommand onExit) {
@@ -64,14 +96,13 @@ public class AvoidingState extends AbstractStateNode {
         super.entry(context);
         context.haltRobot();
         context.moveSensor(0);
+        escapeDir(context).ifPresentOrElse(dir -> put(context, ESCAPE_DIRECTION, dir),
+                () -> remove(context, ESCAPE_DIRECTION));
+        remove(context, FREE_POINT);
     }
 
     @Override
     public String step(ProcessorContext ctx) {
-        if (!isBlocked(ctx)) {
-            // Halt robot at exit
-            return COMPLETED_EXIT;
-        }
         if (isTimeout(ctx)) {
             // Halt robot at timeout
             ctx.haltRobot();
@@ -79,19 +110,67 @@ public class AvoidingState extends AbstractStateNode {
         }
         RobotStatus status = ctx.getRobotStatus();
         int dir = status.getDirection();
-        if (status.canMoveForward()) {
+        if (!status.canMoveForward()) {
+            // Robot blocked forward
+            if (status.canMoveBackward()) {
+                // Robot can move backward
+                // move robot backward
+                int escapeDir = normalizeDegAngle(dir + 180);
+                put(ctx, ESCAPE_DIRECTION, escapeDir);
+                remove(ctx, FREE_POINT);
+                logger.atDebug()
+                        .setMessage("{}: move backward {} DEG")
+                        .addArgument(this::getId)
+                        .addArgument(dir)
+                        .log();
+                ctx.moveRobot(dir, -1);
+                return null;
+            }
+            // Robot completely blocked
+            // holt robot
+            ctx.haltRobot();
+            logger.atWarn().setMessage("{}: Robot blocked").addArgument(this::getId).log();
+            return BLOCKED_EXIT;
+        }
+        if (!status.canMoveBackward()) {
+            // Robot can move forward
             // move robot forward
-            ctx.moveRobot(dir, 1F);
+            put(ctx, ESCAPE_DIRECTION, dir);
+            remove(ctx, FREE_POINT);
+            logger.atDebug()
+                    .setMessage("{}: move forward {} DEG")
+                    .addArgument(this::getId)
+                    .addArgument(dir)
+                    .log();
+            ctx.moveRobot(dir, 1);
             return null;
         }
-        if (status.canMoveBackward()) {
-            // move robot backward
-            ctx.moveRobot(dir, -1F);
-            return null;
+        // Robot not blocked
+        Point2D freePoint = get(ctx, FREE_POINT);
+        int escapeDir = getInt(ctx, ESCAPE_DIRECTION);
+        if (freePoint == null) {
+            freePoint = ctx.getRobotStatus().getLocation();
+            put(ctx, FREE_POINT, freePoint);
+            logger.atDebug().setMessage("{}: escaping to {} DEG from {}")
+                    .addArgument(this::getId)
+                    .addArgument(escapeDir)
+                    .addArgument(freePoint)
+                    .log();
         }
-        // holt robot if blocked
-        ctx.haltRobot();
-        logger.warn("{}: Robot blocked", getId());
+        double distance = freePoint.distance(ctx.getRobotStatus().getLocation());
+        double safeDistance = getDouble(ctx, SAFE_DISTANCE);
+        if (distance >= safeDistance) {
+            // Halt robot at exit
+            logger.atDebug()
+                    .setMessage("{}: safety at {} m")
+                    .addArgument(this::getId)
+                    .addArgument(() -> format("%.2f", distance))
+                    .log();
+            ctx.haltRobot();
+            return COMPLETED_EXIT;
+        }
+        // move robot away
+        ctx.moveRobot(escapeDir, 1);
         return null;
     }
 }