+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue #146: Change explorer agent behavior
 - Issue #148: Update radar map on contacts signals
+- Issue #149: Rotate robot in avoiding behavior
 - Issue #151: Move robot ahead when exploring
 - Issue #152: Sector unknown on partial empty in polar map